requirements.txt
Flask >= 1.0.2, < 1.1.0
flake8-import-order-spoqa >= 1.5.0, < 1.6.0
pycodestyle == 2.3.1
pytest-flake8 >= 1.0.2, < 1.1.0
requests >= 2.19.1 , < 2.20.0
typeguard >= 2.2.2, < 2.3.0
requests_mock >= 1.5.2, < 1.6.0
